Add SQuaRE-style tox and GitHub Actions setup
The SQuaRE approach for testing in GitHub Actions is a bit different
from the Pipelines approach. In SQuaRE we keep library dependencies in
pyproject.toml with a "dev" extra for development dependencies.
Pipelines instead uses requirements.txt and typing.txt requirements
files for development dependencies for compatibility with a number of
reusable workflows. On the other hand, SQuaRE uses a pretty minimal GHA
workflow with the specifics encoded in a tox file that can also be used
locally.

For expedience in getting up and running I'm going to port the SQuaRE
approach here, but we can revert to using the Pipelines reusable
workflows if that's required (though I do worry that there unnecessary
duplication of dependencies and variation of tested dependencies vs
published dependencies).